Fix GPRABBITMQ-38: on reload, start any stopped listeners.
When a developer modifies an artifact that a message listener service depends
on, the message listener is re-registered in the application context. As we
are setting the autoStartup property to false, that means the listeners are
stopped but not restarted. After any class is reloaded, we now check for any
stopped listeners and start them if we find any.